# backend/outputs/ApiContract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PIA Developer Console API",
    "description": "PIA Platform Observability System \u2014 All data served through stable projections and versioned contracts.",
    "version": "2.0.0",
    "x-stability": "frozen"
  },
  "paths": {
...
    "/api/v1/registry/algorithms/{algorithm_id}": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Algorithm",
        "operationId": "get_algorithm_api_v1_registry_algorithms__algorithm_id__get",
        "parameters": [
          {
            "name": "algorithm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Algorithm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registry/projections/{projection_id}": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Projection",
        "operationId": "get_projection_api_v1_registry_projections__projection_id__get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projections/{projection_id}/analytics": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Projection Analytics",
        "description": "Projection Dependency Explorer: return DAG of dependencies for visual rendering.",
        "operationId": "get_projection_analytics_api_v1_registry_projections__projection_id__analytics_get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projections/{projection_id}/dependency-graph": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Projection Dependency Graph",
        "description": "Projection Dependency Explorer: return DAG of dependencies for visual rendering.",
        "operationId": "get_projection_dependency_graph_api_v1_registry_projections__projection_id__dependency_graph_get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/events": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "List Events",
        "description": "Event Explorer: query the immutable event log.",
        "operationId": "list_events_api_v1_registry_events_get",
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Execution Id"
            }
          },
          {
            "name": "repository_session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Repository Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/events/object/{object_id}": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Events For Object",
        "description": "Lineage: all events that touched a specific canonical object.",
        "operationId": "get_events_for_object_api_v1_registry_events_object__object_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registry/projections/{projection_id}/search": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Search Projection",
        "description": "Force a full rebuild of a single projection.",
        "operationId": "search_projection_api_v1_registry_projections__projection_id__search_get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projection-engine/rebuild/{projection_id}": {
      "post": {
        "tags": [
          "Registries"
        ],
        "summary": "Rebuild Projection",
        "description": "Force a full rebuild of a single projection.",
        "operationId": "rebuild_projection_api_v1_registry_projection_engine_rebuild__projection_id__post",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/registry/projection-engine/invalidate/{projection_id}": {
      "post": {
        "tags": [
          "Registries"
        ],
        "summary": "Invalidate Projection",
        "description": "Mark a projection stale without rebuilding it.",
        "operationId": "invalidate_projection_api_v1_registry_projection_engine_invalidate__projection_id__post",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projections/{projection_id}/visualize": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Visualize Projection",
        "description": "Replay from event log and rebuild all affected projections.",
        "operationId": "visualize_projection_api_v1_registry_projections__projection_id__visualize_get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          },
          {
            "name": "focus_node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Focus Node Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projection-engine/replay": {
      "post": {
        "tags": [
          "Registries"
        ],
        "summary": "Replay Projections",
        "description": "Replay from event log and rebuild all affected projections.",
        "operationId": "replay_projections_api_v1_registry_projection_engine_replay_post",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projection-engine/history/{projection_id}": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Projection History",
        "description": "Projection execution history \u2014 every build attempt with metrics.",
        "operationId": "get_projection_history_api_v1_registry_projection_engine_history__projection_id__get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registry/projections/{projection_id}/nodes": {
      "get": {
        "tags": [
          "Registries"
        ],
        "summary": "Get Projection Nodes",
        "description": "Projection execution history \u2014 every build attempt with metrics.",
        "operationId": "get_projection_nodes_api_v1_registry_projections__projection_id__nodes_get",
        "parameters": [
          {
            "name": "projection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Projection Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericObjectDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericObjectDTO": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "GenericObjectDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}